refactor(server): extract unified course-allotment HTML parser and resolve PR review
